backend/requirements-dev.txt
coverage==7.4.0
django-filter-stubs==0.1.3
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
factory_boy==3.3.0
mypy==1.8.0
pylint==3.0.3
ruff==0.4.4
types-pytz==2023.3.1.1
types-requests==2.31.0.20240106
